Create README.md
This repository contains manipulations of digital filters in audio signal processing. It contains four finite impulse response (FIR) filters and one infinite impulse response (IIR) filter projects that aims to contribute to the world of signal and speech processing. 

The projects were created in the form of 'Questions' and 'Solution' with the logic of exercise. Those who want to test themselves can look at the question section first and try to solve the question. You can look at 'Solution' notebooks where you are stuck. Projects structurally determined the intended learning outcome at first. Then, the theory and formulation of the mentioned filter are explained in detail. Then the identified problems are written in a detailed and clean code.